+++ changelog.md
@@ -6,6 +6,10 @@ This is version is compatible with Minecraft 1.21.8 supporting Farmer's Delight 
 
 ***
 
+### Changed
+
+- MidnightLib version updated
+
 ### Fixed
 
 - Fixed Kalan chat messages upon player death (Thank you, bobqianic!)