+++ changelog.md
@@ -9,9 +9,10 @@ NeoForge and Fabric.
 
 ### Changed
 
-- FD 1.2.10 - 1.21.1 Tag Changes
+- FD 1.2.10 - 1.21.1 Tag Changes:
   - All usages of `c:foods/milk` -> `c:drinks/milk`
   - Condensed Milk added to `c:drinks/milk`
+  - Add Ube Cake, Leche Flan, and Leaf Feasts to `c:foods/edible_when_placed`
 - MidnightLib version updated
 
 ### Fixed